+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 CHANGELOG
-=========
+Latest Version: 1.1.0
+=====================
+
+Version 1.1.0 (4/10/2026)
+-------------------------
+    - Added monteCarloSimulations class, and functions to execute simulations
+    - Set functions to static
+    - Latest version added to change log
+    - Changed 'Print' function argument to 'verbose'
+    - Added blackScholes class, and function to execute pricing
+    - Updated TOML file
+    - Test file runs all methods in one file and reports back if they work
+    - Fixed ADX logic
+    - Added calculate greeks class, and functions to claculate the greeks
+    - Notes/Goals:
+        -- Add anti-variance Monte Carlo class
+        -- Make a data set intake Monte Carlo class
+        -- Make a data set intake Black Scholes class
+        -- Change documentation website
+
+Version 1.0.1 (4/10/2026)
+-------------------------
+    - Added test files to ensure scripts work
+    - Base package replaced
+    - Code pushed to GIT
 
 Version 1.0.0 (4/9/2026)
 ------------------------
@@ -13,8 +37,3 @@ Version 1.0.0 (4/9/2026)
     - Notes/Goals:
         -- Base package published (take down ASAP) (V 0.1.0)
         -- Website should be expanded for github links
-
-
-Version 1.0.1 (4/10/2026)
--------------------------
-    - Added test files to ensure scripts work